+++ src/Libs/FShade.GLSL/IntrinsicFunctions.fs
@@ -2077,6 +2077,9 @@ module IntrinsicFunctions =
             CIntrinsic.tagged "ivec3(gl_WorkGroupSize)", [ exactly <@ getWorkGroupSize @> ]
             CIntrinsic.tagged "barrier()", [ exactly <@ barrier @> ]
 
+            CIntrinsic.tagged "SetMeshOutputsEXT({0}, {1})", [ exactly <@ setMeshOutputs @> ]
+            CIntrinsic.tagged "EmitMeshTasksEXT({0}, {1}, {2})", [ exactly <@ emitMeshTasksXYZ @> ]
+
             // ==========================================================================
             // Raytracing
             // ==========================================================================